Deletion of user Profile - authorization, testing done

+++ user/src/main/java/sgsits/cse/dis/user/serviceImpl/userProfileServiceImpl/UserAddressService.java
@@ -7,14 +7,13 @@ import sgsits.cse.dis.user.components.UserProfileRepo;
 import sgsits.cse.dis.user.dtos.UserAddressDto;
 import sgsits.cse.dis.user.dtos.UserProfileDto;
 import sgsits.cse.dis.user.exception.InternalServerError;
-import sgsits.cse.dis.user.model.UserAddress;
+import sgsits.cse.dis.user.exception.UnauthorizedException;
 import sgsits.cse.dis.user.model.UserAddress;
 import sgsits.cse.dis.user.service.UserProfileService;
 
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 @SuppressWarnings("ConstantConditions")
 @Service
@@ -45,34 +44,54 @@ public class UserAddressService implements UserProfileService {
     }
 
     public void addUserProfileElement(final UserAddressDto userAddressDto,
-                                      final String token) throws InternalServerError {
+                                      final String token) throws InternalServerError, UnauthorizedException {
 
         final UserAddress userAddress =
                 userProfileDtoMapper.convertUserAddressDtoIntoModel(userAddressDto);
 
         LOGGER.info("Adding or Updating userElement for id : " + userAddress.getUserId());
 
+        final String userId = jwtResolver.getIdFromJwtToken(token);
+
         if (0 == userAddress.getId()) {
 
-            userAddress.setCreatedBy(jwtResolver.getIdFromJwtToken(token));
+            userAddress.setCreatedBy(userId);
             userAddress.setCreatedDate(new Date(new java.util.Date().getTime()));
-        } else {
+        }  else {
 
             final UserAddress userAddressExisting = userProfileRepo
                     .getUserAddressById(userAddress.getId());
+
+            if(!userAddressExisting.getUserId().equals(userId)) {
+
+                throw new UnauthorizedException("You aren't authorized to edit this");
+            }
+
             userAddress.setCreatedBy(userAddressExisting.getCreatedBy());
             userAddress.setCreatedDate(userAddressExisting.getCreatedDate());
         }
 
-        userAddress.setModifiedBy(jwtResolver.getIdFromJwtToken(token));
+        userAddress.setModifiedBy(userId);
         userAddress.setModifiedDate(new Date(new java.util.Date().getTime()));
 
         userProfileRepo.addOrUpdateUserAddress(userAddress);
     }
 
-    @Override
-    public void deleteUserProfileElementById(final Long id) throws InternalServerError {
 
-        userProfileRepo.deleteUserAddressById(id);
+    @Override
+    public void deleteUserProfileElementById(final Long id, final String token)
+            throws InternalServerError, UnauthorizedException {
+
+        LOGGER.info("Deleting User Address for user id : " + jwtResolver.getIdFromJwtToken(token));
+
+        final UserAddress userAddress =
+                userProfileRepo.getUserAddressById(id);
+
+        if (jwtResolver.getIdFromJwtToken(token).equals(userAddress.getUserId())) {
+
+            userProfileRepo.deleteUserAddressById(id);
+        } else {
+            throw new UnauthorizedException("You are not authorized to delete this");
+        }
     }
 }
