+++ user/src/main/java/sgsits/cse/dis/user/serviceImpl/userProfileServiceImpl/UserProjectService.java
@@ -4,17 +4,14 @@ import org.springframework.stereotype.Service;
 import sgsits.cse.dis.user.components.UserProfileRepo;
 import sgsits.cse.dis.user.dtos.UserProfileDto;
 import sgsits.cse.dis.user.dtos.UserProjectDto;
-import sgsits.cse.dis.user.dtos.UserQualificationDto;
 import sgsits.cse.dis.user.exception.InternalServerError;
+import sgsits.cse.dis.user.exception.UnauthorizedException;
 import sgsits.cse.dis.user.model.UserProject;
-import sgsits.cse.dis.user.model.UserQualification;
-import sgsits.cse.dis.user.model.UserWorkExperience;
 import sgsits.cse.dis.user.service.UserProfileService;
 
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 @Service
 public class UserProjectService implements UserProfileService {
@@ -38,34 +35,55 @@ public class UserProjectService implements UserProfileService {
         return new ArrayList<>(userProjectDtoList);
     }
 
-    public void addUserProfileElement(final UserProjectDto userProjectDto, final String token) throws InternalServerError {
+    public void addUserProfileElement(final UserProjectDto userProjectDto,
+                                      final String token) throws InternalServerError, UnauthorizedException {
 
         final UserProject userProject =
                 userProfileDtoMapper.convertUserProjectDtoIntoModel(userProjectDto);
 
         LOGGER.info("Adding or Updating userElement for id : " + userProject.getUserId());
 
+        final String userId = jwtResolver.getIdFromJwtToken(token);
+
         if (0 == userProject.getId()) {
 
-            userProject.setCreatedBy(jwtResolver.getIdFromJwtToken(token));
+            userProject.setCreatedBy(userId);
             userProject.setCreatedDate(new Date(new java.util.Date().getTime()));
         }  else {
 
             final UserProject userProjectExisting = userProfileRepo
                     .getUserProjectById(userProject.getId());
+
+            if(!userProjectExisting.getUserId().equals(userId)) {
+
+                throw new UnauthorizedException("You aren't authorized to edit this");
+            }
+
             userProject.setCreatedBy(userProjectExisting.getCreatedBy());
             userProject.setCreatedDate(userProjectExisting.getCreatedDate());
         }
 
-        userProject.setModifiedBy(jwtResolver.getIdFromJwtToken(token));
+        userProject.setModifiedBy(userId);
         userProject.setModifiedDate(new Date(new java.util.Date().getTime()));
 
-        userProject.setUserId(jwtResolver.getIdFromJwtToken(token));
         userProfileRepo.addOrUpdateUserProject(userProject);
     }
 
+
     @Override
-    public void deleteUserProfileElementById(final Long id) throws InternalServerError {
-        userProfileRepo.deleteUserProjectById(id);
+    public void deleteUserProfileElementById(final Long id, final String token)
+            throws InternalServerError, UnauthorizedException {
+
+        LOGGER.info("Deleting Project for user id : " + jwtResolver.getIdFromJwtToken(token));
+
+        final UserProject userProject =
+                userProfileRepo.getUserProjectById(id);
+
+        if (jwtResolver.getIdFromJwtToken(token).equals(userProject.getUserId())) {
+
+            userProfileRepo.deleteUserProjectById(id);
+        } else {
+            throw new UnauthorizedException("You are not authorized to delete this");
+        }
     }
 }
